+++ browser.d.ts
@@ -1,3 +1,0 @@
-import loadPikchr = require("./index");
-
-export = loadPikchr;